+++ objc2-foundation/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Added
+* Implement `UnwindSafe` and `RefUnwindSafe` for all objects.
+
 ### Removed
 * `NSObject::hash_code`, `NSObject::is_equal` and `NSObject::description` in
   favour of just having the trait implementations `Hash`, `PartiqalEq` and